requirements.txt
asgiref==3.4.1
Django==3.1.7
django-betterforms==1.2
gunicorn==20.1.0
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
